requirements.txt
aiohttp==3.7.4.post0
asn1crypto==0.24.0
async-timeout==3.0.1
attrs==21.2.0
certifi==2018.8.24
chardet==3.0.4
colorzero==1.1
cryptography==2.6.1
discord.py==1.7.3
entrypoints==0.3
gpiozero==1.5.1
idna==2.6
keyring==17.1.1
keyrings.alt==3.1.1
multidict==5.2.0
pycrypto==2.6.1
PyGObject==3.30.4
python-apt==1.8.4.3
pyxdg==0.25
requests==2.21.0
RPi.GPIO==0.7.0
SecretStorage==2.3.1
six==1.12.0
spidev==3.4
ssh-import-id==5.7
typing-extensions==3.10.0.2
urllib3==1.24.1
yarl==1.7.0
